+++ testing/fmtest.py
@@ -21,31 +21,32 @@ from pymisp import PyMISP
 # -u --url = url/ip address of MISP server
 # -p --port = port server is listening on
 
-def processFile(inputfile,serverurl):
+def processfile(inputfile, serverurl):
     headers = {'content-type': 'application/json'}
     try:
         with open(inputfile) as json_file:
 
             file_data = json_file.read()
+            file_data = file_data.replace("\n", "")
+
             print(json.dumps(file_data, sort_keys=True, indent=4, separators=(',', ': ')))
             try:
                 r = requests.post(serverurl, data=file_data, headers=headers, timeout=5)
-            except Exception, e:
-                print " "
-                print "COMMUNICATION ERROR : " + str(e)
-                print " "
+            except Exception as err:
+
+                print("COMMUNICATION ERROR : " + format(err))
                 sys.exit(2)
-    except Exception, e:
-        print " "
-        print "FILE ERROR : " + str(e)
-        print " "
+    except Exception as err:
+        print("File ERROR : " + format(err))
 
-    print inputfile + " sent to " + serverurl + ". Status code: " + str(r.status_code) + "."
+        return False
 
-    return
+    print (inputfile + " sent to " + serverurl + ". Status code: " + str(r.status_code) + ".")
+
+    return True
 
 def printopts():
-    print '''
+    print ('''
     USAGE:
     -f --file       a specific json file to send
     -d --dir        directory of json files to send. Use ./ for current directory
@@ -58,7 +59,7 @@ def printopts():
     fstest.py -f ./testalert.json -u localhost -p 8080
     fstest.py -d ./alerts -t 2 -u 192.168.1.2 -p 8080
     fstest.py -d ./ -u localhost -p 8888
-    '''
+    ''')
 
 def main(argv):
     inputfile = ''
@@ -75,7 +76,7 @@ def main(argv):
         sys.exit(2)
 
     if not len(opts):
-        print 'No options specified:'
+        print('No options specified:')
         printopts()
         sys.exit(2)
     for opt, arg in opts:
@@ -97,7 +98,7 @@ def main(argv):
 
     # if no url or port --> error
     if (url == '') or (port == ''):
-        print "ERROR: url and port are required"
+        print("ERROR: url and port are required")
         printopts()
         sys.exit(2)
 
@@ -105,26 +106,26 @@ def main(argv):
 
     # go try to read file and send
     if (mode == 'file'):
-        processFile(inputfile,serverurl)
+        processfile(inputfile, serverurl)
     elif (mode == 'directory'):
         filelist = glob.glob(inputdir + '*.json')
         if len(filelist):
             for afile in filelist:
-                processFile(afile,serverurl)
+                processfile(afile, serverurl)
                 time.sleep(float(timeout))
         filelist = glob.glob(inputdir + '*.xml')
         if len(filelist):
             for afile in filelist:
-                processFile(afile, serverurl)
+                processfile(afile, serverurl)
                 time.sleep(float(timeout))
 
 
 
 
         else:
-            print "No files of type .json found in directory: " + inputdir
+            print ("No files of type .json found in directory: " + inputdir)
     else:
-        print "unknown mode"
+        print ("unknown mode")
 
 if __name__ == "__main__":
    main(sys.argv[1:])